+++ src/acetime/zonedbpy/zone_policies.py
@@ -3,11 +3,10 @@
 #   $ /home/brian/src/AceTimeTools/src/acetimetools/tzcompiler.py
 #     --input_dir /home/brian/dev/tz
 #     --output_dir /home/brian/src/AceTimePython/src/acetime/zonedbpy
-#     --tz_version 2021a
+#     --tz_version 2021c
 #     --action zonedb
 #     --language python
 #     --scope extended
-#     --ignore_buf_size_too_large
 #     --start_year 1974
 #     --until_year 2050
 #
@@ -23,13 +22,13 @@
 #   northamerica
 #   southamerica
 #
-# from https://github.com/eggert/tz/releases/tag/2021a
+# from https://github.com/eggert/tz/releases/tag/2021c
 #
 # DO NOT EDIT
 
 #---------------------------------------------------------------------------
-# Supported zone policies: 118
-# numRules: 1142
+# Supported zone policies: 116
+# numRules: 1139
 #---------------------------------------------------------------------------
 
 #---------------------------------------------------------------------------
@@ -1677,67 +1676,19 @@ ZONE_POLICY_Azer = {
 }
 
 #---------------------------------------------------------------------------
-# Policy name: Bahamas
-# Rule count: 3
-#---------------------------------------------------------------------------
-ZONE_RULES_Bahamas = [
-    # Rule    Bahamas    1945    only    -    Oct    17    24:00    0    S
-    {
-        'from_year': 1945,
-        'to_year': 1945,
-        'in_month': 10,
-        'on_day_of_week': 0,
-        'on_day_of_month': 17,
-        'at_seconds': 86400,
-        'at_time_suffix': 'w',
-        'delta_seconds': 0,
-        'letter': 'S',
-    },
-    # Rule    Bahamas    1964    1975    -    Oct    lastSun    2:00    0    S
-    {
-        'from_year': 1964,
-        'to_year': 1975,
-        'in_month': 10,
-        'on_day_of_week': 7,
-        'on_day_of_month': 0,
-        'at_seconds': 7200,
-        'at_time_suffix': 'w',
-        'delta_seconds': 0,
-        'letter': 'S',
-    },
-    # Rule    Bahamas    1964    1975    -    Apr    lastSun    2:00    1:00    D
-    {
-        'from_year': 1964,
-        'to_year': 1975,
-        'in_month': 4,
-        'on_day_of_week': 7,
-        'on_day_of_month': 0,
-        'at_seconds': 7200,
-        'at_time_suffix': 'w',
-        'delta_seconds': 3600,
-        'letter': 'D',
-    },
-
-]
-ZONE_POLICY_Bahamas = {
-    'name': 'Bahamas',
-    'rules': ZONE_RULES_Bahamas
-}
-
-#---------------------------------------------------------------------------
 # Policy name: Barb
 # Rule count: 6
 #---------------------------------------------------------------------------
 ZONE_RULES_Barb = [
-    # Anchor: Rule    Barb    1977    1978    -    Oct    Sun>=1    2:00    0    S
+    # Rule    Barb    1944    only    -    Sep    10    6:00u    0    S
     {
-        'from_year': 0,
-        'to_year': 0,
-        'in_month': 1,
+        'from_year': 1944,
+        'to_year': 1944,
+        'in_month': 9,
         'on_day_of_week': 0,
-        'on_day_of_month': 1,
-        'at_seconds': 0,
-        'at_time_suffix': 'w',
+        'on_day_of_month': 10,
+        'at_seconds': 21600,
+        'at_time_suffix': 'u',
         'delta_seconds': 0,
         'letter': 'S',
     },
@@ -5229,30 +5180,6 @@ ZONE_POLICY_Germany = {
 }
 
 #---------------------------------------------------------------------------
-# Policy name: Ghana
-# Rule count: 1
-#---------------------------------------------------------------------------
-ZONE_RULES_Ghana = [
-    # Rule    Ghana    1951    1956    -    Jan     1    2:00    0    GMT
-    {
-        'from_year': 1951,
-        'to_year': 1956,
-        'in_month': 1,
-        'on_day_of_week': 0,
-        'on_day_of_month': 1,
-        'at_seconds': 7200,
-        'at_time_suffix': 'w',
-        'delta_seconds': 0,
-        'letter': 'GMT',
-    },
-
-]
-ZONE_POLICY_Ghana = {
-    'name': 'Ghana',
-    'rules': ZONE_RULES_Ghana
-}
-
-#---------------------------------------------------------------------------
 # Policy name: Greece
 # Rule count: 12
 #---------------------------------------------------------------------------
@@ -7162,7 +7089,7 @@ ZONE_POLICY_Japan = {
 
 #---------------------------------------------------------------------------
 # Policy name: Jordan
-# Rule count: 32
+# Rule count: 33
 #---------------------------------------------------------------------------
 ZONE_RULES_Jordan = [
     # Anchor: Rule    Jordan    1973    1975    -    Oct    1    0:00    0    -
@@ -7525,10 +7452,10 @@ ZONE_RULES_Jordan = [
         'delta_seconds': 0,
         'letter': '-',
     },
-    # Rule    Jordan    2014    max    -    Mar    lastThu    24:00    1:00    S
+    # Rule    Jordan    2014    2021    -    Mar    lastThu    24:00    1:00    S
     {
         'from_year': 2014,
-        'to_year': 9999,
+        'to_year': 2021,
         'in_month': 3,
         'on_day_of_week': 4,
         'on_day_of_month': 0,
@@ -7548,6 +7475,18 @@ ZONE_RULES_Jordan = [
         'at_time_suffix': 's',
         'delta_seconds': 0,
         'letter': '-',
+    },
+    # Rule    Jordan    2022    max    -    Feb    lastThu    24:00    1:00    S
+    {
+        'from_year': 2022,
+        'to_year': 9999,
+        'in_month': 2,
+        'on_day_of_week': 4,
+        'on_day_of_month': 0,
+        'at_seconds': 86400,
+        'at_time_suffix': 'w',
+        'delta_seconds': 3600,
+        'letter': 'S',
     },
 
 ]
@@ -11461,9 +11400,9 @@ ZONE_POLICY_Poland = {
 # Rule count: 9
 #---------------------------------------------------------------------------
 ZONE_RULES_Port = [
-    # Rule    Port    1951    1965    -    Oct    Sun>=1     2:00s    0    -
+    # Rule    Port    1947    1965    -    Oct    Sun>=1     2:00s    0    -
     {
-        'from_year': 1951,
+        'from_year': 1947,
         'to_year': 1965,
         'in_month': 10,
         'on_day_of_week': 7,
@@ -14329,10 +14268,10 @@ ZONE_RULES_WS = [
         'delta_seconds': 3600,
         'letter': '-',
     },
-    # Rule    WS    2012    max    -    Apr    Sun>=1    4:00    0    -
+    # Rule    WS    2012    2021    -    Apr    Sun>=1    4:00    0    -
     {
         'from_year': 2012,
-        'to_year': 9999,
+        'to_year': 2021,
         'in_month': 4,
         'on_day_of_week': 7,
         'on_day_of_month': 1,
@@ -14341,10 +14280,10 @@ ZONE_RULES_WS = [
         'delta_seconds': 0,
         'letter': '-',
     },
-    # Rule    WS    2012    max    -    Sep    lastSun    3:00    1    -
+    # Rule    WS    2012    2020    -    Sep    lastSun    3:00    1    -
     {
         'from_year': 2012,
-        'to_year': 9999,
+        'to_year': 2020,
         'in_month': 9,
         'on_day_of_week': 7,
         'on_day_of_month': 0,
@@ -15155,130 +15094,6 @@ ZONE_POLICY_Zion = {
 
 
 #---------------------------------------------------------------------------
-
-ZONE_POLICY_MAP = {
-    'AN': ZONE_POLICY_AN,
-    'AQ': ZONE_POLICY_AQ,
-    'AS': ZONE_POLICY_AS,
-    'AT': ZONE_POLICY_AT,
-    'AV': ZONE_POLICY_AV,
-    'AW': ZONE_POLICY_AW,
-    'Albania': ZONE_POLICY_Albania,
-    'Algeria': ZONE_POLICY_Algeria,
-    'Arg': ZONE_POLICY_Arg,
-    'Armenia': ZONE_POLICY_Armenia,
-    'Aus': ZONE_POLICY_Aus,
-    'Austria': ZONE_POLICY_Austria,
-    'Azer': ZONE_POLICY_Azer,
-    'Bahamas': ZONE_POLICY_Bahamas,
-    'Barb': ZONE_POLICY_Barb,
-    'Belgium': ZONE_POLICY_Belgium,
-    'Belize': ZONE_POLICY_Belize,
-    'Bermuda': ZONE_POLICY_Bermuda,
-    'Brazil': ZONE_POLICY_Brazil,
-    'Bulg': ZONE_POLICY_Bulg,
-    'CO': ZONE_POLICY_CO,
-    'CR': ZONE_POLICY_CR,
-    'C_Eur': ZONE_POLICY_C_Eur,
-    'Canada': ZONE_POLICY_Canada,
-    'Chatham': ZONE_POLICY_Chatham,
-    'Chile': ZONE_POLICY_Chile,
-    'Cook': ZONE_POLICY_Cook,
-    'Cuba': ZONE_POLICY_Cuba,
-    'Cyprus': ZONE_POLICY_Cyprus,
-    'Czech': ZONE_POLICY_Czech,
-    'DR': ZONE_POLICY_DR,
-    'Denmark': ZONE_POLICY_Denmark,
-    'Dhaka': ZONE_POLICY_Dhaka,
-    'EU': ZONE_POLICY_EU,
-    'EUAsia': ZONE_POLICY_EUAsia,
-    'E_Eur': ZONE_POLICY_E_Eur,
-    'E_EurAsia': ZONE_POLICY_E_EurAsia,
-    'Ecuador': ZONE_POLICY_Ecuador,
-    'Edm': ZONE_POLICY_Edm,
-    'Egypt': ZONE_POLICY_Egypt,
-    'Eire': ZONE_POLICY_Eire,
-    'Falk': ZONE_POLICY_Falk,
-    'Fiji': ZONE_POLICY_Fiji,
-    'Finland': ZONE_POLICY_Finland,
-    'France': ZONE_POLICY_France,
-    'GB_Eire': ZONE_POLICY_GB_Eire,
-    'Germany': ZONE_POLICY_Germany,
-    'Ghana': ZONE_POLICY_Ghana,
-    'Greece': ZONE_POLICY_Greece,
-    'Guam': ZONE_POLICY_Guam,
-    'Guat': ZONE_POLICY_Guat,
-    'HK': ZONE_POLICY_HK,
-    'Haiti': ZONE_POLICY_Haiti,
-    'Halifax': ZONE_POLICY_Halifax,
-    'Holiday': ZONE_POLICY_Holiday,
-    'Hond': ZONE_POLICY_Hond,
-    'Hungary': ZONE_POLICY_Hungary,
-    'Iran': ZONE_POLICY_Iran,
-    'Iraq': ZONE_POLICY_Iraq,
-    'Italy': ZONE_POLICY_Italy,
-    'Japan': ZONE_POLICY_Japan,
-    'Jordan': ZONE_POLICY_Jordan,
-    'Kyrgyz': ZONE_POLICY_Kyrgyz,
-    'LH': ZONE_POLICY_LH,
-    'Latvia': ZONE_POLICY_Latvia,
-    'Lebanon': ZONE_POLICY_Lebanon,
-    'Libya': ZONE_POLICY_Libya,
-    'Macau': ZONE_POLICY_Macau,
-    'Malta': ZONE_POLICY_Malta,
-    'Mauritius': ZONE_POLICY_Mauritius,
-    'Mexico': ZONE_POLICY_Mexico,
-    'Moldova': ZONE_POLICY_Moldova,
-    'Moncton': ZONE_POLICY_Moncton,
-    'Mongol': ZONE_POLICY_Mongol,
-    'Morocco': ZONE_POLICY_Morocco,
-    'NC': ZONE_POLICY_NC,
-    'NT_YK': ZONE_POLICY_NT_YK,
-    'NZ': ZONE_POLICY_NZ,
-    'Namibia': ZONE_POLICY_Namibia,
-    'Neth': ZONE_POLICY_Neth,
-    'Nic': ZONE_POLICY_Nic,
-    'Norway': ZONE_POLICY_Norway,
-    'PRC': ZONE_POLICY_PRC,
-    'Pakistan': ZONE_POLICY_Pakistan,
-    'Palestine': ZONE_POLICY_Palestine,
-    'Para': ZONE_POLICY_Para,
-    'Peru': ZONE_POLICY_Peru,
-    'Phil': ZONE_POLICY_Phil,
-    'Poland': ZONE_POLICY_Poland,
-    'Port': ZONE_POLICY_Port,
-    'ROK': ZONE_POLICY_ROK,
-    'Romania': ZONE_POLICY_Romania,
-    'Russia': ZONE_POLICY_Russia,
-    'RussiaAsia': ZONE_POLICY_RussiaAsia,
-    'SA': ZONE_POLICY_SA,
-    'Salv': ZONE_POLICY_Salv,
-    'SanLuis': ZONE_POLICY_SanLuis,
-    'Spain': ZONE_POLICY_Spain,
-    'SpainAfrica': ZONE_POLICY_SpainAfrica,
-    'StJohns': ZONE_POLICY_StJohns,
-    'Sudan': ZONE_POLICY_Sudan,
-    'Swiss': ZONE_POLICY_Swiss,
-    'Syria': ZONE_POLICY_Syria,
-    'Taiwan': ZONE_POLICY_Taiwan,
-    'Thule': ZONE_POLICY_Thule,
-    'Tonga': ZONE_POLICY_Tonga,
-    'Toronto': ZONE_POLICY_Toronto,
-    'Troll': ZONE_POLICY_Troll,
-    'Tunisia': ZONE_POLICY_Tunisia,
-    'Turkey': ZONE_POLICY_Turkey,
-    'US': ZONE_POLICY_US,
-    'Uruguay': ZONE_POLICY_Uruguay,
-    'Vanc': ZONE_POLICY_Vanc,
-    'Vanuatu': ZONE_POLICY_Vanuatu,
-    'WS': ZONE_POLICY_WS,
-    'W_Eur': ZONE_POLICY_W_Eur,
-    'Winn': ZONE_POLICY_Winn,
-    'Zion': ZONE_POLICY_Zion,
-
-}
-
-#---------------------------------------------------------------------------
 # Unsupported zone policies: 22
 #---------------------------------------------------------------------------
 
@@ -15307,13 +15122,12 @@ ZONE_POLICY_MAP = {
 
 
 #---------------------------------------------------------------------------
-# Notable zone policies: 10
+# Notable zone policies: 9
 #---------------------------------------------------------------------------
 
 # Belize (["LETTER 'CDT' not single character", "LETTER 'CST' not single character"])
 # DR (["LETTER '-0430' not single character", "LETTER 'EST' not single character"])
 # GB_Eire (["LETTER 'BST' not single character", "LETTER 'GMT' not single character"])
-# Ghana (["LETTER 'GMT' not single character"])
 # Guam (["AT '2:01' not on 15-minute boundary"])
 # Moncton (["AT '0:01' not on 15-minute boundary"])
 # Namibia (["LETTER 'CAT' not single character", "LETTER 'WAT' not single character"])